Authorization complete
Authorization implemented, users have to login/register before accessing the main page. Amins can change user info's and moderators can't.

+++ static/cameras/delete_cameras.js
@@ -1,3 +1,12 @@
+let errorMsg = "";
+
+function validate(number){
+    errorMsg = "";
+    
+    if (!Number(number)){errorMsg = "Id must be a number!"}
+    return Number(number);
+}
+
 function init() {
     const cookies = document.cookie.split("=");
     const token = cookies[cookies.length -1];
@@ -7,11 +16,19 @@ function init() {
 
         let number = document.getElementById("id").value
 
-        document.getElementById("id").value = "";
+        if (!validate(number)){
+            document.getElementById('error').innerHTML = errorMsg;
+            document.getElementById('error').style.color = "red";
+        }
+        else{
+            fetch("http://127.0.0.1:8000/api/cameras/" + number,{
+                method: 'delete',
+                headers: {'Authorization': `Bearer ${token}`}
+            })
+            document.getElementById('error').innerHTML = "Successfully deleted.";
+            document.getElementById('error').style.color = "green";
 
-        fetch("http://127.0.0.1:8000/api/cameras/" + number,{
-            method: 'delete',
-            headers: {'Authorization': `Bearer ${token}`}
-        })
+            document.getElementById("id").value = "";
+        }
     });
 }
